chore(FuseFactory): update fuse factory deployment script
