+++ pkg/deployments/tasks/index.js
@@ -1,6 +1,6 @@
 require("./20220313-periphery");
 require("./20220316-initial-public-adapters");
-require("./20220427-fuse-factory");
 require("./20220429-periphery");
 require("./20220517-long-wsteth-adapter");
 require("./20220518-space-factory");
+require("./20220531-fuse-factory");
